Fix printing of error messages when using bazel-root.
Temporary fix for printing error messages in bazel. When using bazel-root, we
currently present error messages by offset as:

  path/to/file@offset+length:

This change will fix it to show line and column numbers, as:

  path/to/file:line:column:

Soon I'll look into adopting the multi-root approach instead of using
bazel-root, so we should be able to delete this logic when we no longer support
the old frontend in bazel.

Change-Id: I2a44b87815ca85b4958adc2ac9a8d8289536b12f
Reviewed-on: https://dart-review.googlesource.com/49580
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>