[fasta] Make ConstnessEvaluator evaluate potential constness
Potential constness is determined syntactically, based on the shape of
the expression and potential constness of its nodes.  Constants are not
evaluated in any way, and their values and types don't affect potential
constness of expressions.  'Const' is inserted on a constructor
invocation if it is potential const; 'new' is inserted otherwise.

Bug: http://dartbug.com/32737
Change-Id: Ice13803ebe3abf0a0f47cbbff29d13638c2d8e92
Reviewed-on: https://dart-review.googlesource.com/49683
Reviewed-by: Kevin Millikin <kmillikin@google.com>