Only bypass cache when .packages is invalidated
In a86b864fa9b113f5cc4669dc3959d3b7c9761922 the incremental compiler was
updated to always bypass the cache when getting a UrlTranslator to be
able to handle the case where .packages had been updated.

This was driven by Flutter not invalidating .packages.

This has now been introduced in flutter with
https://github.com/flutter/flutter/pull/16467
and we can now limit when we bypass the cache to when the .packages file
has been invalidated.

This CL does just that, and further more adds tests where .packages either
appears or disappears.

Change-Id: I3e2d001a993a59c115bc00fcf714f97094357d35
Reviewed-on: https://dart-review.googlesource.com/50940
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Peter von der Ahé <ahe@google.com>