Add a missing call to greatestClosure
When we infer that a function's return type is the expected one, we
need to make sure to use the greatest closure of that expectation.
Otherwise, the type UnknownType will leak out of type inference if it
occurs in the expectation.

Fixes https://github.com/dart-lang/sdk/issues/32847

Change-Id: I2ce65c6b38d701073e3071f03b8c3d52442c2f86
Reviewed-on: https://dart-review.googlesource.com/52521
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>
Commit-Queue: Kevin Millikin <kmillikin@google.com>