Fix a crash in type inference
It's not generally safe to continue to use the Shadow nodes after they
have had their type inferred, because they give a false picture of the
tree.  In this case it led to an assertion failure.

Move the assert to a place in the code where it is true.  Use the
correct subexpression for subsequent check insertion.  Clear the
parent pointer when Kernel nodes are removed from the tree to try to
catch problems like this.

Fixes https://github.com/dart-lang/sdk/issues/32050

Change-Id: I775747d7a6a5240d63881c8563c83cc84628e3d7
Reviewed-on: https://dart-review.googlesource.com/52525
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>