Fix an incorrect dart:mirrors test
The class defined by an anonymous mixin application is always
abstract.  If it were not, then it would be a Dart 2 static error that
it does not provide implementations of any of the abstract members of
the mixed-in class.  For example:

  class S {}
  abstract class M {
    int f();
  }
  abstract class C extends S with M {}

The class defined by `S with M` must be abstract because it does not
provide an implementation of `f`.

Change-Id: I3a990d7db4c5cb3c2dc83f4f12e4fd1eba68f876
Reviewed-on: https://dart-review.googlesource.com/56107
Commit-Queue: Kevin Millikin <kmillikin@google.com>
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>