Fix status file for analyzer.
It seems like the "missing" analysis error is a strong mode error, so I
don't think the error should even appear without strong mode turned on?

In any case, a "missing" error for non strong users is not something I
would expect to be blocking. Better to fix the build

Change-Id: I5f56b5847b655b501e8146200ab4b84fddbf21cc
Reviewed-on: https://dart-review.googlesource.com/56600
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>